to anydesk

+++ AMPEngine2Lib/Engine/Processes.cpp
@@ -15,10 +15,10 @@ Processes::~Processes(){
 	cfg_all->misc.curIteration = curIteration;
 } // ///////////////////////////////////////////////////////////////////////////
 void Processes::RunAll(){
-	size_t pos = (size_t)rnd.randk();	// max ~ 65500*65500
-	size_t szx = cfg_all->lays.bottomX();
-	int x = int(pos % szx);
-	int y = int(pos / szx);
+	int pos = (int)rnd.randk();	// max ~ 65500*65500
+	int szx = (int)cfg_all->lays.bottomX();
+	int x = pos % szx;
+	int y = pos / szx;
 	int_2 shift = int_2(x, y);
 	processA.RunAll(shift);
